device-simulation/requirements.txt
azure-iot-device==2.12.0
asyncio-mqtt==0.16.2